# external/interaction-shell/backend/openapi.json
{
  "components": {
    "schemas": {
...
      "CaseDiagnosis": {
        "additionalProperties": true,
        "properties": {
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "context_health": {
            "$ref": "#/components/schemas/ContextHealth"
          },
          "evidence_refs": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Refs",
            "type": "array"
          },
          "first_abnormal_step": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Abnormal Step"
          },
          "last_progress_step": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Progress Step"
          },
          "likely_upstream_stage": {
            "title": "Likely Upstream Stage",
            "type": "string"
          },
          "schema_version": {
            "const": "interaction-shell.diagnosis.v1",
            "default": "interaction-shell.diagnosis.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "success": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Success"
          },
          "terminal_code": {
            "title": "Terminal Code",
            "type": "string"
          },
          "terminal_stage": {
            "title": "Terminal Stage",
            "type": "string"
          },
          "trajectory_metrics": {
            "$ref": "#/components/schemas/TrajectoryMetrics"
          }
        },
        "required": [
          "case_id",
          "status",
          "success",
          "terminal_stage",
          "terminal_code",
          "first_abnormal_step",
          "last_progress_step",
          "likely_upstream_stage",
          "confidence",
          "context_health",
          "trajectory_metrics",
          "evidence_refs"
        ],
        "title": "CaseDiagnosis",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Affordance Interaction Shell",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/diagnostics": {
      "get": {
        "operationId": "list_diagnostics_diagnostics_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CaseDiagnosis"
                  },
                  "title": "Response List Diagnostics Diagnostics Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Diagnostics"
      },
      "post": {
        "operationId": "project_diagnostic_diagnostics_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnosisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDiagnosis"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Project Diagnostic"
      }
    },
    "/diagnostics/{case_id}": {
      "get": {
        "operationId": "get_diagnostic_diagnostics__case_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDiagnosis"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Diagnostic"
      }
    },
...
    "/sessions": {
      "post": {
        "operationId": "create_session_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Session"
      }
    },
    "/sessions/{session_id}": {
      "get": {
        "operationId": "get_snapshot_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Session-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeSessionSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Snapshot"
      }
    },
    "/sessions/{session_id}/commands/answer": {
      "post": {
        "operationId": "answer_sessions__session_id__commands_answer_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Session-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerQuestion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "discriminator": {
                    "mapping": {
                      "accepted": "#/components/schemas/Accepted",
                      "conflict": "#/components/schemas/Conflict",
                      "rejected": "#/components/schemas/Rejected",
                      "unsupported": "#/components/schemas/Unsupported"
                    },
                    "propertyName": "kind"
                  },
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/Accepted"
                    },
                    {
                      "$ref": "#/components/schemas/Conflict"
                    },
                    {
                      "$ref": "#/components/schemas/Unsupported"
                    },
                    {
                      "$ref": "#/components/schemas/Rejected"
                    }
                  ],
                  "title": "Response Answer Sessions  Session Id  Commands Answer Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Answer"
      }
    },
    "/sessions/{session_id}/commands/approve": {
      "post": {
        "operationId": "approve_sessions__session_id__commands_approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Session-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "discriminator": {
                    "mapping": {
                      "accepted": "#/components/schemas/Accepted",
                      "conflict": "#/components/schemas/Conflict",
                      "rejected": "#/components/schemas/Rejected",
                      "unsupported": "#/components/schemas/Unsupported"
                    },
                    "propertyName": "kind"
                  },
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/Accepted"
                    },
                    {
                      "$ref": "#/components/schemas/Conflict"
                    },
                    {
                      "$ref": "#/components/schemas/Unsupported"
                    },
                    {
                      "$ref": "#/components/schemas/Rejected"
                    }
                  ],
                  "title": "Response Approve Sessions  Session Id  Commands Approve Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve"
      }
    },
    "/sessions/{session_id}/commands/close": {
      "post": {
        "operationId": "close_sessions__session_id__commands_close_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Session-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseSession"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "discriminator": {
                    "mapping": {
                      "accepted": "#/components/schemas/Accepted",
                      "conflict": "#/components/schemas/Conflict",
                      "rejected": "#/components/schemas/Rejected",
                      "unsupported": "#/components/schemas/Unsupported"
                    },
                    "propertyName": "kind"
                  },
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/Accepted"
                    },
                    {
                      "$ref": "#/components/schemas/Conflict"
                    },
                    {
                      "$ref": "#/components/schemas/Unsupported"
                    },
                    {
                      "$ref": "#/components/schemas/Rejected"
                    }
                  ],
                  "title": "Response Close Sessions  Session Id  Commands Close Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Close"
      }
    },
...
  }
}